Moved threadpool to its own implementation and added a work slab to avoid malloc, added avx2 sgemm and linal ctx for passing threadpool to linalg functions. also added multithreading to other linalg functions

+++ src/logger.c
@@ -156,7 +156,7 @@ void logger_log(log_level level, const char* format, ...)
                     local_time.tm_hour,
                     local_time.tm_min,
                     local_time.tm_sec,
-                    current_time.tv_usec / 1000,
+                    (int)(current_time.tv_usec / 1000),
                     buffer);
         }
 
@@ -170,7 +170,7 @@ void logger_log(log_level level, const char* format, ...)
                     local_time.tm_hour,
                     local_time.tm_min,
                     local_time.tm_sec,
-                    current_time.tv_usec / 1000,
+                    (int)(current_time.tv_usec / 1000),
                     buffer);
         }
     }
